Mac: Use launchd to start the daemon as part of setup
This PR adds a launchd plist file which use socket activation but due to
initial socket file descriptor hold by the launchd process and not by the
`crc daemon` process, first query to daemon api blocks and one way to overcome
to this problem is use something we are doing in the linux to check the file
descriptor and use for golang net listener on that file. In the mac it is tricky
because then we need cgo which will break the cross compilation by using something
like https://github.com/sstephenson/launch_socket_server/blob/master/src/launch/socket.go

In this PR we are avoiding to use of cgo so as part of setup `launchd start` also take
place which start the daemon process and serve the api request.